+++ src/AbstractBlockRenderer.php
@@ -242,25 +242,18 @@ abstract class AbstractBlockRenderer
 		return $this->registered_fields;
 	}
 
-	/**
-	 * Adds notice to backend if the block is deprecated.
-	 *
-	 * Checks registered block array for 'lemon_deprecated'.
-	 *
-	 * @return bool
-	 */
 	protected function maybe_add_deprecation_notice()
 	{
-		if (!isset($this->attributes['wp_lemon']['deprecated'])) {
+
+		if (!isset($this->attributes['deprecated'])) {
 			return false;
 		}
 
-		$deprecation = $this->attributes['wp_lemon']['deprecated'];
-		$message = sprintf(__('This block is deprecated since version %1$s. Please replace this block in favor of %2$s.', 'bulldozer'), $deprecation['since'], $deprecation['use']);
+		$deprecation = $this->attributes['deprecated'];
+		$message = sprintf(__('This block is deprecated since %1$s. Please replace this block in favor of %2$s.', 'bulldozer'), $deprecation['since'], $deprecation['use']);
 		$this->add_notification($message, 'warning');
 		return true;
 	}
-
 
 	/**
 	 * Method to retrieve the block wrapper attributes.