+++ src/Bulldozer.php
@@ -23,7 +23,7 @@ class Bulldozer
 	/**
 	 * Current Bulldozer version.
 	 */
-	const VERSION = '4.4.0';
+	const VERSION = '4.4.1';
 
 	/**
 	 * Active theme object.